Rydde litt i Bekk auth kode
