+++ src/login/bekk/CopyTokenButton.tsx
@@ -1,10 +1,17 @@
 import React from "react";
-import { useGetAccessToken } from "./example2";
 import { Button, useNotification } from "@utdanningsdirektoratet/lisa";
+import { config } from "../../Utils/envUtils";
+import { msalInstance } from "./bekkLogin";
 
 export const CopyTokenButton = () => {
-	const getAccessToken = useGetAccessToken();
 	const showNotification = useNotification();
+
+	async function getAccessToken(): Promise<string> {
+		const result = await msalInstance.acquireTokenSilent({
+			scopes: [config.bekk.scope],
+		});
+		return result.accessToken;
+	}
 
 	function copyToken() {
 		getAccessToken()
